Fix post-draft lookback features.
ListGroupItem isn't a real dom element and doesn't forward refs so we need an inner div.
